refactor(publish): split smoke machinery into local_smoke submodule
The cargo/mod.rs file had grown to ~300 lines with the production publish
path buried below the LocalRegistry scaffolding. Splitting the smoke-test
machinery into its own submodule lets mod.rs (~90 lines) show dispatch +
production paths on one screen.

Along the way, tighten the LocalRegistry abstraction:
- install_binary moves into spawn() so the constructor brings up a fully
  working registry in one call, no caller foot-gun.
- publish_each becomes a method on LocalRegistry instead of a free
  function threading api_url through.
- api_url is set at construction; no String::new() placeholder mutated
  via wait_until_ready. Readiness-failure cleanup happens inline before
  Self exists, so the empty-string intermediate state is gone.
